+++ smart-contracts/contracts/utils/EmptyImpl.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.8.21;
+
+contract EmptyImpl {
+  constructor() {}
+}